+++ GGXrdReversalTool.Library/Scenarios/Event/BlockTypes.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel;
+
+namespace GGXrdReversalTool.Library.Scenarios.Event
+{
+    public enum BlockTypes
+    {
+        Normal,
+        [Description("Instant Block")]
+        InstantBlock,
+        [Description("Faultless Defense")]
+        FaultlessDefense,
+        Random
+    }
+}
