+++ GGXrdReversalTool/Converters/BlockTypeToStringConverter.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+using System.Windows.Data;
+using GGXrdReversalTool.Library.Scenarios.Event;
+
+namespace GGXrdReversalTool.Converters;
+
+[ValueConversion(typeof(BlockTypes), typeof(string))]
+[ValueConversion(typeof(IEnumerable<BlockTypes>), typeof(IEnumerable<string>))]
+public class BlockTypeToStringConverter : EnumToStringConverter<BlockTypes> { }
